Add an implementation that saves the Athena query history to S3
It traverses the history backwards and saves the executions in batches of 10K in gzipped files on S3, and it stores the first processed query execution ID in a state file. On startup it reads the state file and then stops processing when it finds the query execution ID from the last run.

It formats timestamps to be compatible with Athena (i.e. no ISO8601, that would be too easy), and adds the region to the objects. It writes the files in a non-Hive partitioned structure, mostly to keep some order. The data will most likely be too small to require partitioning in Athena.

The code is meant to be run in Lambda on a schedule.