Change the S3 keys for the query logs to be less deeply nested
Even with tens of thousands of queries run per day the amount of data is not huge. With a deep directory structure Athena will have to do a lot of S3 listings, which hurts performance a lot. Athena doesn't really understand S3 very well, and will treat it as a file system. With a deep directory structure it will do a breadth first listing of all "directories".

The idea with the key format was that we might want to partition the tables eventually. A variant of the old scheme that looks more reasonable is to use Hive style partitioning (because why not, and Athena likes it better) with only the region and the month (as the first date of the month).